Spawning rocks and cans.

+++ Ported/joe-jia-ryan/Assets/ECS/Components/ThrowingArmsSharedDataComponent.cs
@@ -3,7 +3,7 @@ using Unity.Entities;
 [System.Serializable]
 public struct ThrowingArmsSharedDataComponent : ISharedComponentData
 {
-    public float ArmCount;
+    public int ArmCount;
     public float ArmWidth;
     public float ConveyorMargin;
     public float ConveyorWidth;
